requirements.txt
celery==5.5.3
fastapi==0.115.4
pynest-api==0.3.2
playwright==1.54.0
python-dotenv==1.1.1
requests==2.32.4
uvicorn==0.32.0
beautifulsoup4==4.13.4
lxml==6.0.0
redis==6.4.0
